ceate seperate route for chats. and redirecting on chats or landing page based on authentication status
